refactor(LottoGenerator): purchaseAmount의 검증 코드를 InputView에서 LottoGenerator로 변경
InputView는 Input만 받는 역할로, LottoGenerator에 메세지를 던져 검증하는게 맞다고 생각
